requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
PyMuPDF==1.23.3
openai==0.27.4
python-dotenv==1.0.0
supabase==0.1.1
logfire==0.1.3
httpx==0.24.1